refactor(is): refining structure for focus identity exceptions
