+++ docs/api/is/focusable.md
@@ -28,6 +28,13 @@ var isFocusable = ally.is.focusable(element);
 | ---- | ---- | ------- | ----------- |
 | element | [`HTMLElement`](https://developer.mozilla.org/en/docs/Web/API/HTMLElement) | *required* | The Element to test. |
 
+The underlying rules can also be accessed in the [`options` argument style](../concepts.md#Single-options-argument) by calling `ally.is.focusable.rules(options)`:
+
+| Name | Type | Default | Description |
+| ---- | ---- | ------- | ----------- |
+| context | [`<selector>`](../concepts.md#Selector) | *required* | The element to examine. The first element of a collection is used. |
+| except | [`<focus identification exception>`](../concepts.md#Focus-identification-exceptions) | `{}` | The Element to test. |
+
 ### Returns
 
 Boolean, `true` if the element is focusable.
@@ -43,6 +50,7 @@ Boolean, `true` if the element is focusable.
 ## Changes
 
 * Since `v#master` all `<area>` elements are considered focus-relevant, but only valid `<area>` elements are considered focusable.
+* Since `v#master` exceptions can be passed to `ally.is.focusable.rules(options)`.
 
 
 ## Notes
