+++ test/unit/is.only-tabbable.test.js
@@ -39,7 +39,13 @@ define([
           isOnlyTabbable(null);
         }).to.throw(TypeError, 'is/only-tabbable requires an argument of type Element');
       },
-
+      '.rules() and .except()': function() {
+        var element = document.getElementById('inert-div');
+        expect(isOnlyTabbable.rules({
+          context: element,
+        })).to.equal(false, '.rules()');
+        expect(isOnlyTabbable.rules.except({})(element)).to.equal(false, '.rules.except()');
+      },
       'label with tabindex="-1"': function() {
         var element = document.getElementById('label-tabindex--1');
         expect(isOnlyTabbable(element)).to.equal(false);
